# specs/001-school-library-system/contracts/api-spec.yaml
openapi: 3.0.3
info:
  title: BCD Library Management System API
  description: |
    REST API for school library management (Bibliothèque Centre Documentaire).

    **Operational Model**: Librarian-operated system with offline-first design.
    - No fines/fees (privilege blocking instead)
    - No email/SMS notifications (printed reports)
    - Librarian-mediated access (no student self-service)

    **Key Features**:
    - Circulation (checkout/return/renew with barcode scanning)
    - Cataloging (BNF SRU API integration for ISBN lookup)
    - Borrower management (students/teachers/staff organized by class)
    - Holds/reservations (librarian-mediated queue management)
    - Reports (overdue, statistics, never-borrowed)
    - Barcode generation (Code 39/128)
  version: 1.0.0
  contact:
    name: BCD API Support
    email: support@bcd-library.local
  license:
    name: MIT
servers:
  - url: http://localhost:8000/api/v1
    description: Development server (localhost)
  - url: http://library-server.local:8000/api/v1
    description: Production server (local network)
...
paths:
...
  /holds:
    post:
      tags: [holds]
      summary: Place a hold (librarian-mediated)
      description: |
        Librarian places a hold on behalf of a borrower for a bibliographic record.
        Assigns queue position automatically.
      operationId: createHold
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - borrower_id
                - bibliographic_record_id
              properties:
                borrower_id:
                  type: string
                  example: "105"
                bibliographic_record_id:
                  type: integer
                  example: 1
                notes:
                  type: string
                  example: "Student very interested in this series"
      responses:
        '201':
          description: Hold successfully created
          content:
            application/json:
              schema:
                type: object
                properties:
                  hold_id:
                    type: integer
                  queue_position:
                    type: integer
                  estimated_availability:
                    type: string
                    format: date
                    nullable: true
                  borrower:
                    type: object
                    properties:
                      borrower_id:
                        type: string
                      full_name:
                        type: string
                  bibliographic_record:
                    type: object
                    properties:
                      id:
                        type: integer
                      title:
                        type: string
                      total_items:
                        type: integer
                      items_available:
                        type: integer
              example:
                hold_id: 5
                queue_position: 2
                estimated_availability: "2026-02-20"
                borrower:
                  borrower_id: "105"
                  full_name: "Emma BERNARD"
                bibliographic_record:
                  id: 1
                  title: "Ils ont arrêté mon père"
                  total_items: 2
                  items_available: 0
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /holds/{hold_id}:
    delete:
      tags: [holds]
      summary: Cancel a hold
      description: Cancels a hold and reorders the queue
      operationId: cancelHold
      parameters:
        - name: hold_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Hold cancelled successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
        '404':
          $ref: '#/components/responses/NotFound'
  /holds/bibliographic/{bibliographic_record_id}:
    get:
      tags: [holds]
      summary: Get holds for a title
      description: Returns all active holds for a bibliographic record with queue positions
      operationId: getHoldsForTitle
      parameters:
        - name: bibliographic_record_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: List of holds for this title
          content:
            application/json:
              schema:
                type: object
                properties:
                  bibliographic_record:
                    type: object
                    properties:
                      id:
                        type: integer
                      title:
                        type: string
                  holds:
                    type: array
                    items:
                      type: object
                      properties:
                        hold_id:
                          type: integer
                        borrower_id:
                          type: string
                        borrower_name:
                          type: string
                        class_name:
                          type: string
                        hold_date:
                          type: string
                          format: date-time
                        queue_position:
                          type: integer
                        status:
                          type: string
        '404':
          $ref: '#/components/responses/NotFound'
  /holds/borrower/{borrower_id}:
    get:
      tags: [holds]
      summary: Get borrower's holds
      description: Returns all active holds for a borrower
      operationId: getBorrowerHolds
      parameters:
        - name: borrower_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Borrower's active holds
          content:
            application/json:
              schema:
                type: object
                properties:
                  borrower:
                    $ref: '#/components/schemas/BorrowerSummary'
                  holds:
                    type: array
                    items:
                      type: object
                      properties:
                        hold_id:
                          type: integer
                        bibliographic_record_id:
                          type: integer
                        title:
                          type: string
                        queue_position:
                          type: integer
                        status:
                          type: string
                        hold_date:
                          type: string
                          format: date-time
                        estimated_availability:
                          type: string
                          format: date
                          nullable: true
        '404':
          $ref: '#/components/responses/NotFound'
...
  /reports/overdue:
    get:
      tags: [reports]
      summary: Generate overdue items report
      description: |
        Returns overdue items organized by class for printing.
        Optionally filter by specific class.
      operationId: getOverdueReport
      parameters:
        - name: class_id
          in: query
          schema:
            type: integer
        - name: format
          in: query
          schema:
            type: string
            enum: [json, pdf]
            default: json
      responses:
        '200':
          description: Overdue report
          content:
            application/json:
              schema:
                type: object
                properties:
                  generated_at:
                    type: string
                    format: date-time
                  classes:
                    type: array
                    items:
                      type: object
                      properties:
                        class_name:
                          type: string
                        grade_level:
                          type: string
                        overdue_items:
                          type: array
                          items:
                            type: object
                            properties:
                              borrower_name:
                                type: string
                              item_id:
                                type: string
                              title:
                                type: string
                              due_date:
                                type: string
                                format: date
                              days_overdue:
                                type: integer
            application/pdf:
              schema:
                type: string
                format: binary
...
  /reports/most-borrowed:
    get:
      tags: [reports]
      summary: Get most borrowed titles
      operationId: getMostBorrowedReport
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
        - name: period
          in: query
          description: Time period for statistics
          schema:
            type: string
            enum: [week, month, year, all-time]
            default: year
      responses:
        '200':
          description: Most borrowed titles
          content:
            application/json:
              schema:
                type: object
                properties:
                  period:
                    type: string
                  bibliographic_records:
                    type: array
                    items:
                      type: object
                      properties:
                        bibliographic_record:
                          $ref: '#/components/schemas/BiblographicRecord'
                        circulation_count:
                          type: integer

  # ========================================
  # ADMIN ENDPOINTS
  # ========================================
...
components:
  schemas:
    BiblographicRecord:
      type: object
      properties:
        id:
          type: integer
        isbn:
          type: string
          nullable: true
        title:
          type: string
        subtitle:
          type: string
          nullable: true
        authors:
          type: array
          items:
            type: string
        illustrators:
          type: array
          items:
            type: string
        publisher:
          type: string
          nullable: true
        publication_year:
          type: integer
          nullable: true
        collection:
          type: string
          nullable: true
        series_number:
          type: string
          nullable: true
        language:
          type: string
          nullable: true
          description: ISO 639 language code
        country_code:
          type: string
          nullable: true
        binding_type:
          type: string
          nullable: true
          enum: [hardcover, paperback, spiral, other]
        category:
          type: string
          nullable: true
        genre:
          type: string
          nullable: true
        level:
          type: string
          nullable: true
        medium_type:
          type: string
          enum: [Livre, CD, DVD, Revue, Magazine, Autre]
        target_audience:
          type: string
          nullable: true
          enum: [child, youth, adult]
        keywords:
          type: array
          items:
            type: string
        description:
          type: string
          nullable: true
        page_count:
          type: integer
          nullable: true
        has_illustrations:
          type: boolean
        dimensions:
          type: string
          nullable: true
        total_items:
          type: integer
        total_circulations:
          type: integer
...
    BorrowerSummary:
      type: object
      properties:
        borrower_id:
          type: string
        full_name:
          type: string
        class_name:
          type: string
          nullable: true
        current_loans:
          type: integer
        loan_limit:
          type: integer
        active:
          type: boolean
        blocked_reason:
          type: string
          nullable: true
...
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
          nullable: true
  responses:
    BadRequest:
      description: Bad request (validation error)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "validation_error"
            message: "Invalid request parameters"
            details:
              field: "borrower_id"
              issue: "Required field missing"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "not_found"
            message: "Resource not found"